requirements.txt
pandas
numpy
python-dotenv
pymongo[srv]
pymongo
certifi
scikit-learn
matplotlib
seaborn
apache-airflow

# -e .